+++ CHANGES.rst
@@ -30,6 +30,7 @@ Development
 - Add singleton factory method to ``TerkinDatalogger``. Thanks, @DieDiren.
 - Increase timeout for ARP ping requests with "terkin.py"
 - Use most recent "dotty_dict" module
+- Properly format MAC addresses
 
 
 2019-06-22 0.5.1